Breakpoints file should be temporary
Using vim's built in functionality for getting a temp file so we don't
need to manage it ourselfs. This reduces the overall complexity of the
plugin with the trade off of not letting the user customize the path to
save the file. Since the file doesn't need to be persisted across
executions of vim there is no real need to use anything other then a
temp file.